requirements.txt
aws-cdk-lib==2.66.0
aws-cdk.aws-sagemaker-alpha==2.66.0a0
constructs>=10.0.0,<11.0.0
